# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignIn": {
        "properties": {
          "condition_code": {
            "title": "Condition Code",
            "type": "string"
          },
          "learner_email": {
            "title": "Learner Email",
            "type": "string"
          }
        },
        "required": [
          "learner_email",
          "condition_code"
        ],
        "title": "AssignIn",
        "type": "object"
      },
...
      "ChunkDetail": {
        "properties": {
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "heading_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading Path"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "document_id",
          "filename",
          "text",
          "heading_path",
          "page_start",
          "page_end"
        ],
        "title": "ChunkDetail",
        "type": "object"
      },
...
      "ConditionIn": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "config": {
            "additionalProperties": true,
            "default": {},
            "title": "Config",
            "type": "object"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        },
        "required": [
          "code",
          "description"
        ],
        "title": "ConditionIn",
        "type": "object"
      },
...
      "ExperimentCreateIn": {
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionIn"
            },
            "title": "Conditions",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "conditions"
        ],
        "title": "ExperimentCreateIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetrievalOut": {
        "properties": {
          "char_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char End"
          },
          "char_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Start"
          },
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "heading_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading Path"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "document_id",
          "text",
          "heading_path",
          "content_type",
          "page_start",
          "page_end",
          "char_start",
          "char_end",
          "score",
          "source"
        ],
        "title": "RetrievalOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/chunks/{chunk_id}": {
      "get": {
        "description": "The source-viewer's read: open one cited chunk by id.",
        "operationId": "get_chunk_api_v1_courses__course_id__chunks__chunk_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chunk_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chunk Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chunk",
        "tags": [
          "retrieval"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/retrieval": {
      "get": {
        "description": "Query chunks scoped to one course the caller owns.\n\nfrozen-scope.md: retrieval is restricted to documents attached to the\ncurrent learner's current course. The ownership filter lives inside\nRetrievalService's queries, not here -- this route only translates the\nresult to JSON.",
        "operationId": "retrieve_api_v1_courses__course_id__retrieval_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 1000,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RetrievalOut"
                  },
                  "title": "Response Retrieve Api V1 Courses  Course Id  Retrieval Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve",
        "tags": [
          "retrieval"
        ]
      }
    },
...
    "/api/v1/evaluation/experiments": {
      "post": {
        "operationId": "create_experiment_api_v1_evaluation_experiments_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluation/experiments/{experiment_id}/assign": {
      "post": {
        "operationId": "assign_learner_api_v1_evaluation_experiments__experiment_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Learner",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluation/experiments/{experiment_id}/conditions": {
      "get": {
        "operationId": "list_conditions_api_v1_evaluation_experiments__experiment_id__conditions_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Conditions",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/evaluation/experiments/{experiment_id}/my-condition": {
      "get": {
        "description": "A learner checking which condition they're in for an experiment they\nwere assigned to -- not gated to the experiment's owner, since the\ncaller is only ever reading their own assignment (user.id, never a\nquery param).",
        "operationId": "get_my_condition_api_v1_evaluation_experiments__experiment_id__my_condition_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-email",
            "required": true,
            "schema": {
              "title": "X-User-Email",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": true,
            "schema": {
              "title": "X-Internal-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Condition",
        "tags": [
          "evaluation"
        ]
      }
    },
...
  }
}